chore: move util function to shared file
